Do not allow direct visit to invest page if it disabled
